Added handling for the case that a TMY EPW file for a WMO already exists, so that we don't download it again if we already have it

+++ lib/diyepw/diyepw/get_tmy_epw_file.py
@@ -27,7 +27,6 @@ def get_tmy_epw_file(wmo_index:int, output_dir:str = None):
 
     if output_dir is None:
         output_dir = _os.path.join(_files_dir, "tmy_epw_files")
-        _logger.debug(f"get_tmy_epw_file() - output_dir was not defined - TMY EPW will be written to {output_dir}")
 
     # The sources we know of for TMY EPW files are http://climate.onebuilding.org and https://energyplus.net/weather;
     # we use the climate.onebuilding.org source here because it has all of the EPW files linked
@@ -39,7 +38,7 @@ def get_tmy_epw_file(wmo_index:int, output_dir:str = None):
     # a set of TMY files for every WMO in the United States
     global _catalog_html
     if _catalog_html is None:
-        _logger.info(f"get_tmy_epw_file() - We don't have the catalog of EPW files for WMO {wmo_index} yet. Downloading it from {catalog_url}")
+        _logger.info(f"get_tmy_epw_file() - We don't have the catalog of EPW files for the USA yet. Downloading it from {catalog_url}")
 
         # Retrieve the TMY EPW catalog for the requested year.
         try:
@@ -53,21 +52,27 @@ def get_tmy_epw_file(wmo_index:int, output_dir:str = None):
     if match is None:
         raise Exception(f"No file for WMO index {wmo_index} could be found at {catalog_url}")
     file_url = catalog_url + match.groups()[0]
+    epw_file_name = file_url.split("/")[-1].replace(".zip", ".epw")
+    epw_file_path = _os.path.join(output_dir, epw_file_name)
+
+    # Check whether we already have the TMY EPW file in our directory. If we do it would be a waste of time
+    # to download it again.
+    if _os.path.exists(epw_file_path):
+        _logger.debug(f"get_tmy_epw_file() - TMY EPW file ({epw_file_path}) already exists, won't download again.")
+        return epw_file_path
 
     # Download the ZIP file and decompress it. It contains a number of files including the EPW that we are looking for.
     tmp_file_handle, tmp_file_path = _tempfile.mkstemp()
     tmp_dir = _tempfile.mkdtemp()
-    _logger.debug(f"get_tmy_epw_file() - Downloading file from {file_url}")
+    _logger.debug(f"get_tmy_epw_file() - Downloading file from {file_url} and saving to {epw_file_path}")
     with _request.urlopen(file_url) as response:
         with open(tmp_file_handle, 'wb') as downloaded_file:
             downloaded_file.write(response.read())
     with _ZipFile(tmp_file_path, 'r') as zip_file:
         zip_file.extractall(tmp_dir)
 
-    # Move the EPW file from the temporary directory into which we extracted the ZIP file into the directory storing
-    # our EPWs
-    epw_file_name = file_url.split("/")[-1].replace(".zip", ".epw")
-    epw_file_path = _os.path.join(output_dir, epw_file_name)
+    # Move the EPW file from the temporary directory into which we extracted
+    # the ZIP file into the directory storing our EPWs
     _os.rename(_os.path.join(tmp_dir, epw_file_name), epw_file_path)
 
     # Delete the temporary files created in this call
